Se deja funcionando el agregar cantidad al carrito desde campo numerico en el detalle del producto

+++ dominio/Articulo.cs
@@ -25,7 +25,7 @@ namespace dominio
         public string PrecioFormateado => Precio.ToString("F2");
         public bool Estado { get; set; }
 
-        public float Stock {  get; set; } 
+        public int Stock {  get; set; } 
 
         public Articulo()
         {
